+++ pkg/server/ingest_test.go
@@ -119,7 +119,7 @@ var _ = Describe("server", func() {
 
 						reg := prometheus.NewRegistry()
 
-						s, err := storage.New(storage.NewConfig(&(*cfg).Server), logrus.StandardLogger(), reg, new(health.Controller))
+						s, err := storage.New(storage.NewConfig(&(*cfg).Server), logrus.StandardLogger(), reg, new(health.Controller), storage.NoopApplicationMetadataService{})
 						Expect(err).ToNot(HaveOccurred())
 						defer s.Close()
 						e, _ := exporter.NewExporter(nil, nil)